server/embeddings/requirements.txt
flask
chromadb
sentence-transformers
requests
python-dotenv
tqdm
